console.log("hello world");
JavaScript is used to make Websites dynamic and interactive. Its mostly run on web browsers 
but here is a weapon named as Console (used to test code and fix bugs because errors 
are not visible on browser).
Bugs: Errors in software to produce incorrect or unexpected results.
Syntax is given above with example of hello World.